requirements.txt
flask
python-dotenv
elevenlabs
langchain-community
streamlit
langchain-groq